backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
numpy>=1.26.0
scipy>=1.11.0
pillow>=10.1.0
python-multipart>=0.0.6
pytest>=7.4.0
ruff>=0.1.0
